+++ examples/tree_2d_dgsem/elixir_mhd_ec.jl
@@ -9,8 +9,8 @@ equations = IdealGlmMhdEquations2D(1.4)
 
 initial_condition = initial_condition_weak_blast_wave
 
-volume_flux = (flux_hindenlang, flux_nonconservative_powell)
-solver = DGSEM(polydeg=3, surface_flux=(flux_hindenlang, flux_nonconservative_powell),
+volume_flux = (flux_hindenlang_gassner, flux_nonconservative_powell)
+solver = DGSEM(polydeg=3, surface_flux=(flux_hindenlang_gassner, flux_nonconservative_powell),
                volume_integral=VolumeIntegralFluxDifferencing(volume_flux))
 
 coordinates_min = (-2, -2)